+++ CHANGELOG.md
@@ -4,6 +4,10 @@ Since v0.0.10, this Changelog is formatted according to the [Common Changelog][c
 
 ## UNRELEASED
 
+### Fixed
+
+- ([#](https://github.com/demos-europe/demosplan-ui/pull/)) Add missing prop clearOnSelect to DpMultiselect ([@spiess-demos](https://github.com/spiess-demos))
+
 ## v0.3.5 - 2024-01-09
 
 ### Fixed